refactor: extract collection guards and delay utility from content.ts
Move DEFAULT_COLLECTION_GUARDS, GUARD_FIELD_NAMES,
GUARD_ARRAY_FIELD_NAMES, loadCollectionGuards,
parseGuardFieldNames, applyCollectionGuards into new
lib/collection-guards.ts module. Move delay() into
lib/dom-utils.ts as an exported utility.

content.ts: 1687→1630 lines (-57)
